% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_ts}
\alias{df_to_ts}
\title{Convert a data frame in a time series}
\usage{
df_to_ts(df)
}
\arguments{
\item{df}{}
}
\value{
time series
}
\description{
Convert a data frame in a time series
}
\examples{
1+1
}
