% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_yaxis.R
\name{max_yaxis}
\alias{max_yaxis}
\title{Find the maximum drawing in the Y axis}
\usage{
max_yaxis(list_stacked, list_unstacked)
}
\arguments{
\item{list_stacked}{Dygraph lists that is intended to show stacked}

\item{list_unstacked}{Dygraph lists that is intended to show unstacked}
}
\value{
a double value
}
\description{
Find the maximum drawing in the Y axis
}
\examples{
1+1
}
