% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_back.R
\name{viz_back_potential}
\alias{viz_back_potential}
\title{Visualize backshift}
\usage{
viz_back_potential(obj, show_fixed = TRUE, show_cap = TRUE,
  zoom_cap = FALSE, stacked = TRUE, aggregate = c("none", "object",
  "flex", "all"))
}
\arguments{
\item{obj}{an `e_frame` object}

\item{show_fixed}{display the original fixed demand}

\item{aggregate}{level of aggregation.}
}
\value{
dygraph with the results
}
\description{
Visualize backshift
}
\examples{
1+1
}
