% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_edit.R
\name{viz_bundle}
\alias{viz_bundle}
\title{Bundle several dygraphs}
\usage{
viz_bundle(..., ymax = NULL, names = NULL)
}
\arguments{
\item{...}{dygraphs}

\item{ymax}{if supplied, the y randge of all graphs is et between 0 and `ymax`}

\item{names}{names to give to the graphs}
}
\value{
list of graphs
}
\description{
Bundle several dygraphs
}
\examples{
1+1
}
