% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_fore.R
\name{viz_fore_input}
\alias{viz_fore_input}
\title{Visualize foreshift}
\usage{
viz_fore_input(obj, show_fixed = TRUE, aggregate = c("none", "object",
  "flex", "all"))
}
\arguments{
\item{obj}{an `e_frame` object}

\item{show_fixed}{display the original fixed demand}
}
\value{
dygraph with the results
}
\description{
Visualize foreshift
}
\examples{
1+1
}
