% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_specifics.R
\name{viz_production_fixed}
\alias{viz_production_fixed}
\title{Compare two or more dygraphs}
\usage{
viz_production_fixed(obj)
}
\arguments{
\item{obj}{e_frame object.}
}
\value{
A dygraph
}
\description{
Compare two or more dygraphs
}
\examples{
1+1
}
