% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_storage.R
\name{viz_storage_flows}
\alias{viz_storage_flows}
\title{Visualize backshift}
\usage{
viz_storage_flows(obj)
}
\arguments{
\item{obj}{an `e_frame` object}

\item{show_fixed}{display the original fixed demand}

\item{aggregate}{level of aggregation.}
}
\value{
dygraph with the results
}
\description{
Visualize backshift
}
\examples{
1+1
}
