% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizget.R
\name{vizget}
\alias{vizget}
\title{Visualization gadget for bundles of graphs}
\usage{
vizget(dybundle)
}
\arguments{
\item{dybundle}{Object generated with viz_bundle}
}
\value{
Shiny gadget
}
\description{
Visualization gadget for bundles of graphs
}
\examples{
1+1
}
