% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_yaxis.R
\name{set_yaxis}
\alias{set_yaxis}
\title{Especify the randge of the Y axis in a dygraph}
\usage{
set_yaxis(dy, n)
}
\arguments{
\item{n}{}
}
\value{
A dygraph, with the especification of the Y axis modified
}
\description{
Especify the randge of the Y axis in a dygraph
}
\examples{
1+1
}
