% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_compare.R
\name{viz_compare}
\alias{viz_compare}
\title{Compare two or more dygraphs}
\usage{
viz_compare(dylist, colnames)
}
\arguments{
\item{colnames}{vector with the names to display in dygraph legend}

\item{list}{List of dygraphs to compare. Normally two.}
}
\value{
A dygraph
}
\description{
Compare two or more dygraphs
}
\examples{
1+1
}
